♻️ 43ZECH backend: decompose connect command
