+++ .agentplane/tasks/202605290321-43ZECH/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290321-43ZECH`
+Title: Backend command decomposition
+Canonical task record: `.agentplane/tasks/202605290321-43ZECH/README.md`
+
+## Summary
+
+Backend command decomposition
+
+Decompose packages/agentplane/src/commands/backend.ts by extracting backend connect configuration and dotenv helpers into a focused module, preserving CLI behavior while bringing backend.ts under the 400-line hotspot warning threshold.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/backend.ts by extracting backend connect configuration and dotenv helpers into a focused module, preserving CLI behavior while bringing backend.ts under the 400-line hotspot warning threshold.
+- Out of scope: unrelated refactors not required for "Backend command decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:21:56.833Z
+- Branch: task/202605290321-43ZECH/backend-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
