+++ .agentplane/tasks/202605290321-43ZECH/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 43ZECH task: Backend command decomposition [202605290321-43ZECH]
